refactor: Line 클래스 분리 - 한 줄 랜덤 생성 로직과 재가공하는 로직 분리
