+++ docs/README.md
@@ -1,3 +1,38 @@
+## 기능 구현 목록
+
+### Model
+- [x] 사다리 라인(`|-----|`)
+  - [x] `true` 가로 작대기 있음
+  - [x] `false` 가로 작대기 없음
+- [x] 사다리 maker
+  - [x] Boolean을 한 줄 랜덤하게 생성
+  - [x] 랜덤 생성한 Boolean line을 가로라인이 이웃끼리 겹치지 않도록 조정
+  - [x] 사다리 높이 예외처리(0을 제외한 숫자만 가능)
+- [x] 사람
+  - [x] 입력조건 예외처리
+  - [x] 이름 리스트를 형식에 맞게 공백 넣어주기
+
+### View
+#### 입력 조건
+- [x] 사람 이름 입력하기
+  - [x] 최대 5글자 (사람 이름이 5글자 이하라면 공백문자로 채워줘야 함)
+  - [x] 쉼표 기준으로 입력받기
+  - [x] 첫 번째 사람인 경우에, 이름이 5글자 이하여도 여백을 공백문자로 채우지 않음
+- [x] 사다리 높이 입력받기
+  - [x] 숫자만 입력받기 (공백 or 문자가 입력될 경우 예외처리)
+
+#### 출력 조건
+- [x] 이름이 5글자일 경우, 이름의 끝 글자를 기준으로 입력 받은 높이만큼 사다리의 세로 라인 출력
+- [x] 이름이 5글자 이하인 경우, 이름의 끝 글자 + 공백문자를 기준으로 입력 받은 높이만큼 사다리의 세로 라인 출력
+- [x] 사다리가 정상적으로 작동하려면 가로 라인이 겹치지 않게 하기(ex.`|-----|-----|`)
+- x] 사다리 라인(`|-----|`)
+  - [x] `true` -> `-----|`
+  - [x] `false` -> `     |`
+
+### Controller
+- [x] 사다리 타리 진행
+
+---
 # 요구사항
 사다리 타기 미션 저장소
 - 사다리 게임에 참여하는 사람의 이름을 최대 5글자까지 부여할 수 있다. 사다리를 출력할 때 사람 이름도 같이 출력한다.
@@ -17,31 +52,4 @@
 - 모든 원시 값과 문자열을 포장한다
 - 줄여 쓰지 않는다(축약 금지).
 - 일급 컬렉션을 쓴다.
-
-## 기능 구현 목록
-
-### Model
-- [x] 사다리(`|-----|`)
-  - [x] 막대기 세로라인
-  - [x] 막대기 가로라인
-- [x] 사다리 maker
-  - [x] 가로라인이 이웃 겹치지 않도록 랜덤하게 생성
-  - [x] 사다리 높이 예외처리(0을 제외한 숫자만 가능)
-- [x] 사람
-  - [x] 입력조건 예외처리
-
-### View
-#### 입력 조건
-- [x] 사람 이름 입력하기
-  - [x] 최대 5글자 (사람 이름이 5글자 이하라면 공백문자로 채워줘야 함)
-  - [x] 쉼표 기준
-  - [x] 첫 번째 사람인 경우에, 이름이 5글자 이하여도 여백을 공백문자로 채우지 않음
-- [x] 사다리 높이 입력받기
-
-#### 출력 조건
-- [x] 이름이 5글자일 경우, 이름의 끝 글자를 기준으로 입력 받은 높이만큼 사다리의 세로 라인 출력
-- [x] 이름이 5글자 이하인 경우, 이름의 끝 글자 + 공백문자를 기준으로 입력 받은 높이만큼 사다리의 세로 라인 출력
-- [x] 사다리가 정상적으로 작동하려면 가로 라인이 겹치지 않게 하기(ex.`|-----|-----|`)
-
-### Controller
-- [x] 사다리 타리 진행 
+-
